cloud_function/requirements.txt
functions-framework==3.8.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
pandas==2.2.2
pyarrow==17.0.0